// ST.UI/src/app/Models/ProposalDetailsVM.ts
export class ProposalDetailsVM
{
    ProposalId: number;
    SaleId: number;
    BidPrice: number;
    BidUnit: number;
    ShareDescription: string;
    NumberOfShares: number;
    UnitPrice: number;
    CreatedDt: Date;
    CreatedBy: number;
    UpdatedDt: Date;
    UpdatedBy: number;
    Status: number;
    StatusComments: string;
    StatusDescription: string;
    IsActive: boolean;
    CreatedBy_String: string;
    UpdatedBy_String:string;
    ShareholderContactNumber: string;
    Seller: string;
    Buyer:string;
    ShareHolderShares: number;
    ValidTo: Date;
}